requirements.txt
PyQt6
moviepy>=1.0.3
Pillow
opencv-python
numpy>=1.26.4
proglog
decorator
imageio
imageio-ffmpeg
tqdm
pyinstaller>=6.3.0 